+++ miden/src/cli/prove.rs
@@ -55,14 +55,14 @@ pub struct ProveCmd {
     security: String,
 
     /// Enable tracing to monitor execution of the VM
-    #[clap(short = 't', long = "tracing")]
-    tracing: bool,
+    #[clap(short = 't', long = "trace")]
+    trace: bool,
 }
 
 impl ProveCmd {
     pub fn get_proof_options(&self) -> Result<ProvingOptions, ExecutionOptionsError> {
         let exec_options =
-            ExecutionOptions::new(Some(self.max_cycles), self.expected_cycles, self.tracing)?;
+            ExecutionOptions::new(Some(self.max_cycles), self.expected_cycles, self.trace, false)?;
         Ok(match self.security.as_str() {
             "96bits" => {
                 if self.rpx {
@@ -145,7 +145,7 @@ fn load_data(params: &ProveCmd) -> Result<(Program, InputFile), Report> {
 
     // load program from file and compile
     let program =
-        ProgramFile::read(&params.assembly_file)?.compile(&Debug::Off, &libraries.libraries)?;
+        ProgramFile::read(&params.assembly_file)?.compile(Debug::Off, &libraries.libraries)?;
 
     // load input data from file
     let input_data = InputFile::read(&params.input_file, &params.assembly_file)?;