+++ processor/src/operations/sys_ops.rs
@@ -1,3 +1,5 @@
+use vm_core::Operation;
+
 use super::{
     super::{
         system::{FMP_MAX, FMP_MIN},
@@ -106,6 +108,18 @@ where
         let clk = self.system.clk();
         self.stack.set(0, Felt::from(clk));
         self.stack.shift_right(0);
+        Ok(())
+    }
+
+    // EVENTS
+    // --------------------------------------------------------------------------------------------
+
+    /// Forwards the emitted event id to the host.
+    pub(super) fn op_emit(&mut self, event_id: u32) -> Result<(), ExecutionError> {
+        self.stack.copy_state(0);
+        self.decoder.set_user_op_helpers(Operation::Emit(event_id), &[event_id.into()]);
+        self.host.borrow_mut().on_event(self, event_id)?;
+
         Ok(())
     }
 }